requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
groq==1.1.2
mistralai==2.4.0
google-genai==1.73.1